Polishing up multi-song

+++ src/modipulate/mod_stream.h
@@ -219,7 +219,7 @@ private:
     friend void mod_stream_callback_finished(void* userData);
 
     // Global volume, from 0.0 to 1.0
-    static double modipulate_global_volume;
+    static float modipulate_global_volume;
 };
 
 #endif // MODSTREAM_H
